refactor selector gui into mvc
